+++ src/producer.rs
@@ -10,7 +10,7 @@ use crate::{
     error::{Error, Result},
     metadata::ClusterMetadata,
     network::BrokerConnection,
-    protocol::{ProduceRequest, ProduceResponse},
+    protocol::{Header, ProduceRequest, ProduceResponse},
     DEFAULT_CLIENT_ID, DEFAULT_CORRELATION_ID,
 };
 
@@ -56,6 +56,7 @@ impl ProduceParams {
 ///         partition_id,
 ///         key: Some(bytes::Bytes::from_static(b"Tester")),
 ///         value: Some(bytes::Bytes::from_static(b"Value")),
+///         headers: vec![String::from("Key"), bytes::Bytes::from("Value")]
 ///     };
 ///
 /// let producer_client = samsa::prelude::ProducerBuilder::new(bootstrap_addrs, vec![topic_name.to_string()])
@@ -82,6 +83,7 @@ pub struct ProducerSink;
 pub struct ProduceMessage {
     pub key: Option<Bytes>,
     pub value: Option<Bytes>,
+    pub headers: Vec<Header>,
     pub topic: String,
     pub partition_id: i32,
 }
@@ -156,6 +158,7 @@ pub async fn produce(
             message.partition_id,
             message.key.clone(),
             message.value.clone(),
+            message.headers.clone(),
         );
     }